backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.27.0
python-dotenv==1.0.1
python-socketio==5.11.0
pydantic[email]==2.8.0
pydantic-settings==2.4.0
ddtrace==2.10.0
datadog==0.49.0
